# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music | API",
    "description": "A RESTful API that works as the core of SNM (Social Network for Music).",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Enter the JWT with the 'Bearer ' prefix (e.g. 'Bearer [token]')."
    }
  },
  "paths": {
...
    "/api/playlists/search/": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Allows to search public playlists",
        "description": "",
        "parameters": [
          {
            "name": "title",
            "description": "Returns only playlists that match the following title",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tag",
            "description": "Returns only playlists that contain the following tag",
            "in": "query",
            "type": "string"
          },
          {
            "name": "track",
            "description": "Returns only playlists that contain the following track",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlaylistPreview"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Returns the current user's playlists",
        "description": "",
        "parameters": [
          {
            "name": "select",
            "description": "Allows to filter by category [all | owner | follower]",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlaylistPreview"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "playlist"
        ],
        "summary": "Creates a new playlist",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Midnight tunes vol.2"
                },
                "description": {
                  "type": "string",
                  "example": "All the best night tunes to play at 00:00am!"
                },
                "public": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "65df6fe0aef53429367c76c9"
                },
                "owner": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "65a3105e29c6516ab2dd1b38"
                    },
                    "username": {
                      "type": "string",
                      "example": "davi0k"
                    }
                  }
                },
                "title": {
                  "type": "string",
                  "example": "Midnight tunes vol.2"
                },
                "description": {
                  "type": "string",
                  "example": "All the best night tunes to play at 00:00am!"
                },
                "public": {
                  "type": "boolean",
                  "example": true
                },
                "tags": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "tracks": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "followers": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Retrieves a playlist's full view",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Updates a current user's playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Midnight tunes vol.2"
                },
                "description": {
                  "type": "string",
                  "example": "All the best night tunes to play at 00:00am!"
                },
                "tags": {
                  "type": "array",
                  "example": [
                    "night"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "playlist"
        ],
        "summary": "Deletes a current user's playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/thumbnail/": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Retrieves a playlist's thumbnail",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "sizes": {
                  "type": "object",
                  "properties": {
                    "64": {
                      "type": "array",
                      "example": [
                        "https://i.scdn.co/image/ab67616d00004851f3aa0e6ca22a382007f61e4d",
                        "https://i.scdn.co/image/ab67616d00004851fff2cb485c36a6d8f639bdba",
                        "https://i.scdn.co/image/ab67616d00004851b98afa12c212cbbda4f1799b"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "300": {
                      "type": "array",
                      "example": [
                        "https://i.scdn.co/image/ab67616d00001e02f3aa0e6ca22a382007f61e4d",
                        "https://i.scdn.co/image/ab67616d00001e02fff2cb485c36a6d8f639bdba",
                        "https://i.scdn.co/image/ab67616d00001e02b98afa12c212cbbda4f1799b"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "640": {
                      "type": "array",
                      "example": [
                        "https://i.scdn.co/image/ab67616d0000b273f3aa0e6ca22a382007f61e4d",
                        "https://i.scdn.co/image/ab67616d0000b273fff2cb485c36a6d8f639bdba",
                        "https://i.scdn.co/image/ab67616d0000b273b98afa12c212cbbda4f1799b"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/add/{track}/": {
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Adds a track to a playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          },
          {
            "name": "track",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A track's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/remove/{track}/": {
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Removes a track from a playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          },
          {
            "name": "track",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A track's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/follow/": {
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Follows a public playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/playlists/{id}/unfollow/": {
      "patch": {
        "tags": [
          "playlist"
        ],
        "summary": "Unfollows a public playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A playlist's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "PlaylistPreview": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "65df6fe0aef53429367c76c9"
        },
        "owner": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "65a3105e29c6516ab2dd1b38"
            },
            "username": {
              "type": "string",
              "example": "gordon.freeman"
            }
          }
        },
        "title": {
          "type": "string",
          "example": "City 17's Tunes"
        },
        "public": {
          "type": "boolean",
          "example": true
        },
        "tags": {
          "type": "array",
          "example": [
            "city",
            "apocalyptic",
            "dystopia"
          ],
          "items": {
            "type": "string"
          }
        },
        "followers": {
          "type": "array",
          "example": [
            "65df84edaef53429367c7c77",
            "65df90c4aef53429367c7d7f"
          ],
          "items": {
            "type": "string"
          }
        },
        "totalTracks": {
          "type": "number",
          "example": 3
        }
      }
    },
...
  }
}